[no-ticket] split custom ID tests, add RUN_IT_SINGLE env var
